fixed error/warnings from clang++

+++ Plugin.Kuwahara/GpuKuwaharaFilterNode.cpp
@@ -132,7 +132,7 @@ public:
 
     bool postInit() override
     {
-        string opts = string_format("-DN_SECTORS=%d", _N);
+        string opts = string_format("-DN_SECTORS=%d", (int)_N);
 
         _kidGeneralizedKuwahara = _gpuComputeModule->registerKernel("generalizedKuwahara", "kuwahara.cl", opts);
         _kidGeneralizedKuwaharaRgb = _gpuComputeModule->registerKernel("generalizedKuwaharaRgb", "kuwahara.cl", opts);
@@ -237,7 +237,7 @@ public:
 
     bool postInit() override
     {
-        string opts = string_format("-DN_SECTORS=%d", _N);
+        string opts = string_format("-DN_SECTORS=%d", (int)_N);
 
         _kidAnisotropicKuwahara = _gpuComputeModule->registerKernel("anisotropicKuwahara", "kuwahara.cl", opts);
         _kidAnisotropicKuwaharaRgb = _gpuComputeModule->registerKernel("anisotropicKuwaharaRgb", "kuwahara.cl", opts);
